+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 This is the changelog for this fork of OMI.
 It documents the changes in each of the tagged releases
 
+## 2.2.1 - 2021-07-13
+
++ Added support for RHEL systems.
+
 ## 2.2.0 - 2021-04-07
 
 + Created universal builds to be used across the various nix distributions